Surface list damage extraction for proper per candidate damage
Previously, candidate damage was computed by looping over all the
|surface_damage_rect_list| in the case where the quad did not have a
|surface_damage_index|. This was both incorrect (excessive damage) and slow (O(m*n)).

Now with our |OverlayCandidateFactory| we are able to find all the
damage that is not associated with a specific quad and store it in a
precomputed rect. It is this damage that we will use when a candidate
quad does not have a |surface_damage_index|.

Bug: 1286058
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3615994
Commit-Queue: Peter McNeeley <petermcneeley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002197}

+++ components/viz/service/display/overlay_processor_using_strategy.cc
@@ -413,7 +413,7 @@ void OverlayProcessorUsingStrategy::UpdateDamageRect(
     const auto& status = it.second;
     if (status.plane_z_order != 0) {
       RecordOverlayDamageRectHistograms(status.plane_z_order > 0,
-                                        status.damage_area_estimate != 0,
+                                        status.damage_area_estimate != 0.f,
                                         damage_rect.IsEmpty());
     }
   }
@@ -496,11 +496,11 @@ void OverlayProcessorUsingStrategy::SortProposedOverlayCandidatesPrioritized(
     // for low latency surfaces (inking like in the google keeps application).
     const bool force_update = it->candidate.overlay_damage_index !=
                                   OverlayCandidate::kInvalidDamageIndex &&
-                              it->candidate.damage_area_estimate != 0;
-    track_data.AddRecord(
-        frame_sequence_number_,
-        static_cast<float>(it->candidate.damage_area_estimate) / display_area,
-        it->candidate.resource_id, tracker_config_, force_update);
+                              it->candidate.damage_area_estimate != 0.f;
+    track_data.AddRecord(frame_sequence_number_,
+                         it->candidate.damage_area_estimate / display_area,
+                         it->candidate.resource_id, tracker_config_,
+                         force_update);
     // Here a series of criteria are considered for wholesale rejection of a
     // candidate. The rational for rejection is usually power improvements but
     // this can indirectly reallocate limited overlay resources to another
